requirements.txt
arabic-reshaper==3.0.0
asn1crypto==1.5.1
bottle==0.12.25
bottle-mysql==0.3.1
certifi==2024.8.30
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cryptography==43.0.1
cssselect2==0.7.0
cycler==0.12.1
fonttools==4.53.1
h5py==3.11.0
html5lib==1.1
idna==3.8
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
mysqlclient==2.2.4
numpy==2.1.1
oscrypto==1.3.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyMySQL==1.1.1
pyparsing==3.1.4
pypdf==4.3.1
pypng==0.20220715.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-ldap==3.4.4
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
scipy==1.14.1
simplekml==1.3.6
six==1.16.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
uuid==1.30
webencodings==0.5.1
xhtml2pdf==0.2.16
